+++ src/ble/sm.c
@@ -1028,17 +1028,19 @@ void sm_cmac_signed_write_start(const sm_key_t k, uint8_t opcode, hci_con_handle
 
 static void sm_trigger_user_response_basic(sm_connection_t * sm_conn, uint8_t event_type){
     setup->sm_user_response = SM_USER_RESPONSE_PENDING;
-    uint8_t event[11];
+    uint8_t event[12];
     sm_setup_event_base(event, sizeof(event), event_type, sm_conn->sm_handle, sm_conn->sm_peer_addr_type, sm_conn->sm_peer_address);
+    event[11] = setup->sm_use_secure_connections ? 1 : 0;
     sm_dispatch_event(HCI_EVENT_PACKET, 0, event, sizeof(event));
 }
 
 static void sm_trigger_user_response_passkey(sm_connection_t * sm_conn){
-    uint8_t event[15];
+    uint8_t event[16];
     uint32_t passkey = big_endian_read_32(setup->sm_tk, 12);
     sm_setup_event_base(event, sizeof(event), SM_EVENT_PASSKEY_DISPLAY_NUMBER, sm_conn->sm_handle,
                         sm_conn->sm_peer_addr_type, sm_conn->sm_peer_address);
-    little_endian_store_32(event, 11, passkey);
+    event[11] = setup->sm_use_secure_connections ? 1 : 0;
+    little_endian_store_32(event, 12, passkey);
     sm_dispatch_event(HCI_EVENT_PACKET, 0, event, sizeof(event));
 }